refactor(controller): extract parser.ts + fix inverted dep (R5)
